Add --host config.sh option and CROSS_COMPILE Makefile variable. (#402)
Adds `--host` to config.sh and its corresponding variable `CROSS_COMPILE` to the build system (previously only used by the Pandora port). This is mainly for Linux, but it also simplified the toolchain variables for several architectures. This required changing how the console ports override `STRIP` and moving `--strip-unneeded` from that variable.

Additionally, I fixed these warnings:
* 3DS, NDS: `as` would repeatedly complain about missing end-of-file newlines caused by `bin2s`.
* NDS: fixed `#ifdef` used inside of the macro `info()`.
* PSP: C++98 whining about `long long` and commas at the end of enum declarations.

+++ src/configure.h
@@ -24,6 +24,8 @@
 
 __M_BEGIN_DECLS
 
+#include <stdint.h>
+
 #define OPTION_NAME_LEN 33
 
 enum config_type
@@ -148,8 +150,8 @@ struct config_info
   // Virtual filesystem options
   boolean vfs_enable;
   boolean vfs_enable_auto_cache;
-  long long vfs_max_cache_size;
-  long long vfs_max_cache_file_size;
+  int64_t vfs_max_cache_size;
+  int64_t vfs_max_cache_file_size;
 
   // Game options
   char startup_path[256];
